查询数据

Created-by: insightDeveloper

Match-id-a546e5116366df10f40d98550a06df9df9b3b0e9

+++ src/protocol/ProtocolRequest.h
@@ -124,7 +124,7 @@ struct ThreadDetailRequest : public Request {
 
 struct UnitFlowNameParams {
     std::string rankId;
-    uint32_t tid = 0;
+    int64_t tid = 0;
     std::string pid;
     uint64_t startTime;
 };
